+++ src/server/game/Globals/ObjectMgr.cpp
@@ -7353,7 +7353,7 @@ AreaTriggerStruct const* ObjectMgr::GetGoBackTrigger(uint32 Map) const
         if (itr->second.target_mapId == entrance_map)
         {
             AreaTriggerEntry const* atEntry = sAreaTriggerStore.LookupEntry(itr->first);
-            if (atEntry && atEntry->ContinentID == Map)
+            if (atEntry && uint32(atEntry->ContinentID) == Map)
                 return &itr->second;
         }
     }
